Move Tree's doc to tree module
